# Optimize modeler.ReadAccessor for sparse accessors